+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Changelog
 
-## v 0.7.4.3 (beta)
+## v 0.7.5
 
 * Export DSL results to gsheets: `DslDataset.save_gsheets()`
 * CLI command: `.export_as_gsheets`
 * Bug fix: json_normalize errors when creating 
 * Updated DSL grammar for dimcli CLI autocomplete
 * fix bug caused by new `_copyright` in JSON response 
-
+* `to_json_file` replaces `DslDataset.save_json`
+* `to_gsheets` replaces `DslDataset.save_gsheets`
+* `utils.export_as_gsheets` added
 
 ## v 0.7.4